endpoint: move status-related functions to `status.go`
Signed-off by: Ian Vernon <ian@cilium.io>
